task 1 complete

+++ src/DataTypes.java
@@ -3,7 +3,7 @@ import java.util.List;
 public class DataTypes {
     public static long sum(List<Integer> numbers) {
 
-        int s = 0;
+        long s = 0;
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
